remove diet selector
