+++ data/configs_for_docs/default_config.yml
@@ -32,4 +32,4 @@ pipeline:
     max_ngram: 4
   - name: DIETClassifier
   - name: EntitySynonymMapper
-  - name: DIETSelector
+  - name: ResponseSelector
