+++ data/configs_for_docs/default_english_config.yml
@@ -12,4 +12,4 @@ pipeline:
     max_ngram: 4
   - name: DIETClassifier
   - name: EntitySynonymMapper
-  - name: DIETSelector
+  - name: ResponseSelector
