+++ docker/configs/config_pretrained_embeddings_convert.yml
@@ -12,4 +12,4 @@ pipeline:
     max_ngram: 4
   - name: DIETClassifier
   - name: EntitySynonymMapper
-  - name: DIETSelector
+  - name: ResponseSelector
