+++ docker/configs/config_pretrained_embeddings_spacy_en.yml
@@ -13,4 +13,4 @@ pipeline:
     max_ngram: 4
   - name: DIETClassifier
   - name: EntitySynonymMapper
-  - name: DIETSelector
+  - name: ResponseSelector
