+++ docker/configs/config_supervised_embeddings_duckling.yml
@@ -11,6 +11,6 @@ pipeline:
     max_ngram: 4
   - name: DIETClassifier
   - name: EntitySynonymMapper
-  - name: DIETSelector
+  - name: ResponseSelector
   - name: DucklingHTTPExtractor
     url: "http://duckling:8000"
