+++ rasa/cli/initial_project/config.yml
@@ -12,7 +12,7 @@ pipeline:
     max_ngram: 4
   - name: DIETClassifier
   - name: EntitySynonymMapper
-  - name: DIETSelector
+  - name: ResponseSelector
 
 # Configuration for Rasa Core.
 # https://rasa.com/docs/rasa/core/policies/
