+++ tests/nlu/test_train.py
@@ -41,7 +41,6 @@ def pipelines_for_tests():
                 "DIETClassifier",
                 "EmbeddingIntentClassifier",
                 "ResponseSelector",
-                "DIETSelector",
                 "EntitySynonymMapper",
             ),
         ),
@@ -60,7 +59,6 @@ def pipelines_for_tests():
                 "SklearnIntentClassifier",
                 "DIETClassifier",
                 "ResponseSelector",
-                "DIETSelector",
                 "EntitySynonymMapper",
             ),
         ),
@@ -77,7 +75,6 @@ def pipelines_for_tests():
                 "DucklingHTTPExtractor",
                 "DIETClassifier",
                 "ResponseSelector",
-                "DIETSelector",
                 "EntitySynonymMapper",
             ),
         ),
@@ -93,7 +90,6 @@ def pipelines_for_tests():
                 "DucklingHTTPExtractor",
                 "DIETClassifier",
                 "ResponseSelector",
-                "DIETSelector",
                 "EntitySynonymMapper",
             ),
         ),
@@ -110,7 +106,6 @@ def pipelines_for_tests():
                 "MitieIntentClassifier",
                 "DIETClassifier",
                 "ResponseSelector",
-                "DIETSelector",
                 "EntitySynonymMapper",
             ),
         ),
@@ -126,7 +121,6 @@ def pipelines_for_tests():
                 "MitieIntentClassifier",
                 "DIETClassifier",
                 "ResponseSelector",
-                "DIETSelector",
                 "EntitySynonymMapper",
             ),
         ),
