+++ tests/utilities.py
@@ -4,7 +4,6 @@ from yarl import URL
 import rasa.utils.io as io_utils
 from nlu.classifiers.diet_classifier import DIETClassifier
 from nlu.classifiers.embedding_intent_classifier import EmbeddingIntentClassifier
-from nlu.selectors.diet_selector import DIETSelector
 from nlu.selectors.response_selector import ResponseSelector
 from utils.tensorflow.constants import EPOCHS
 
@@ -32,7 +31,6 @@ def update_number_of_epochs(config_path: Text, output_file: Text):
             EmbeddingIntentClassifier.name,
             DIETClassifier.name,
             ResponseSelector.name,
-            DIETSelector.name,
         ]:
             component[EPOCHS] = 2
